[hla_plugin]: ajout du CallSignResolver et remaniement de la gestion des identifiants HLA.

+++ src/libraries/hla_plugin/AggregateEntity.cpp
@@ -27,13 +27,13 @@ using namespace plugins::hla;
 // Name: AggregateEntity constructor
 // Created: SBO 2008-02-18
 // -----------------------------------------------------------------------------
-AggregateEntity::AggregateEntity( Agent_ABC& agent, unsigned short identifier,
+AggregateEntity::AggregateEntity( Agent_ABC& agent, unsigned int identifier,
                                   const std::string& name, rpr::ForceIdentifier force, const rpr::EntityType& type )
     : agent_     ( agent )
     , attributes_( new AttributesSerializer() )
 {
     attributes_->Register( "EntityType", type );
-    attributes_->Register( "EntityIdentifier", rpr::EntityIdentifier( 1, 1, identifier ) );
+    attributes_->Register( "EntityIdentifier", rpr::EntityIdentifier( 1, 1, static_cast< unsigned short >( identifier ) ) );
     attributes_->Register( "ForceIdentifier", Wrapper< unsigned char >( static_cast< unsigned char >( force ) ) );
     attributes_->Register( "AggregateMarking", AggregateMarking( name ) );
     attributes_->Register( "AggregateState", Wrapper< unsigned char >( 1 ) ); // fully aggregated
